+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.1
+
+* Fix oauth token request
+* Add example web service for CLI authorization
+
 # 1.0.0
 
 * Initial version